Some minor changes

+++ Epam Task1 deadline 20.05/src/for_all_tasks/Validation.java
@@ -1,12 +1,12 @@
 package for_all_tasks;
 
 public class Validation {
+	
 	/**
 	 * @param n
 	 * @return true if n is negative number
 	 */
-	public static boolean isNegative(int n) {
+	public static boolean isNegative(double n) {
 		return (n < 0);
 	}
-
 }
